+++ src/Settings.h
@@ -41,6 +41,7 @@ public:
         bool extraRingModeChanged {false};
         bool enchantmentStrengthChanged {false};
         bool fingerSelectionChanged {false};
+        bool bondOfMatrimonyLeftRingFingerChanged {false};
         bool npcBondOfMatrimonyLeftRingFingerPreferenceChanged {false};
         bool npcSupportChanged {false};
         bool npcSupportEnabled {true};
@@ -52,6 +53,7 @@ public:
             return extraRingModeChanged
                    || enchantmentStrengthChanged
                    || fingerSelectionChanged
+                   || bondOfMatrimonyLeftRingFingerChanged
                    || npcBondOfMatrimonyLeftRingFingerPreferenceChanged
                    || npcSupportChanged
                    || unequipAllClearsExtraRingsChanged
@@ -69,6 +71,7 @@ public:
     [[nodiscard]] std::uint32_t GetFingerSelectModifierKey() const;
     [[nodiscard]] std::uint32_t GetFingerSelectModifierButton() const;
     [[nodiscard]] bool IsNpcSupportEnabled() const;
+    [[nodiscard]] bool ShouldUseBondOfMatrimonyOnLeftRingFinger() const;
     [[nodiscard]] bool ShouldNpcPreferBondOfMatrimonyOnLeftRingFinger() const;
     [[nodiscard]] bool IsActorVirtualRingSupportEnabled(Core::ActorKey a_actor) const;
     [[nodiscard]] bool ShouldUnequipAllClearExtraRings() const;
@@ -84,6 +87,7 @@ private:
     std::atomic<std::uint32_t> fingerSelectModifierKey_ {kDefaultFingerSelectModifierKey};
     std::atomic<std::uint32_t> fingerSelectModifierButton_ {kDefaultFingerSelectModifierButton};
     std::atomic_bool npcSupportEnabled_ {true};
+    std::atomic_bool bondOfMatrimonyOnLeftRingFinger_ {false};
     std::atomic_bool npcPreferBondOfMatrimonyOnLeftRingFinger_ {true};
     std::atomic_bool unequipAllClearsExtraRings_ {true};
     std::atomic<std::uint16_t> enabledVirtualTargetBits_ {kDefaultEnabledVirtualTargetBits};